requirements.txt
asgiref==3.5.0
boto3==1.23.0
botocore==1.26.0
distlib==0.3.4
Django==4.0.3
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.1.0
platformdirs==2.5.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
whitenoise==6.1.0
